requirements.txt
python-dotenv
openai
langchain
tiktoken
langchain-openai
langchain-community
pandas
tqdm
# transformers  - HuggingFace local
